requirements.txt
sqlalchemy==2.0.46
fastapi==0.129.0
uvicorn==0.41.0
pydantic-settings==2.13.0
bcrypt==5.0.0
pyjwt==2.11.0
aiosqlite==0.22.1